Animation Montage Setup
Montage Created from Rifle Animation. Upper Body Slot assigned to the montage default Group. Man_AnimBP Eventgraph now checks to see if the player Has Nozzle and if so it appropriately sets it's own boolean assigned tot he Anim Graph. In the Anim Graph two new caches are saved. One for normal locomotion, the other for the added blending of the rifle animation montage. in the default slot node connected between normal locomotion and firehose holding locomotion, in the slot name it should be assigned to Upper Body. But for some reason Animation Montages aren't being shown in the asset browser available to be chosen from. Therefore it never gets assigned. Referenced Tutorials are below...

https://youtu.be/y7iLItzwGTM?si=VPw01wrUyKTMFnoB (good to pickup some of the added reasons for the certain nodes used)
https://youtu.be/SxXI3um8gRM?si=z9fyTS3dghZyUwJw   (used the blend by bool logic from this but goes over the similar values as the one before)
https://youtu.be/bU95pf5hTkI?si=Q84lw7JMaYGGmTja  (main tutorial referenced)